Commit session after alembic updates have finished
After alembic updates we need to add foundations. The tables to add
foundations to are created as part of dbupdate process. For some reason
Alembic doesn't commit a session in the end. Because the session is not
commited, the tables do not really get created. And because of that,
foundations fail.

The solution is to commit the session after alembic updates.

(cherry picked from commit a5dae0dbc2dc7036a117589555e3a5b0bf10f596)